requirements.txt
streamlit
requests
python-dateutil
dateparser
faiss-cpu
numpy
groq
python-dotenv
